fix: revert `Avatar` TypeScript conversion
After this landed we were getting some silent playroom webpack svg failures in CI.
